+++ source_shared/protocol.h
@@ -43,7 +43,7 @@ char *ALL_CLIENTS_CONNECTED;
 
 
 void construct_header(char *buffer, struct node node_to_send);
-int receive_message(int client_socket, void *buf, size_t total_bytes_to_receive);
-int send_message(int client_socket, void *buf, size_t total_bytes_to_send);
+ssize_t receive_message(int client_socket, void *buf, ssize_t total_bytes_to_receive);
+ssize_t send_message(int client_socket, void *buf, ssize_t total_bytes_to_send);
 
 #endif
